app-crypt/*: delete all but Cloudflare and DigitalOcean plugins
Since the app-crypt/certbot-3.2.0-r100 ebuild, almost all (except
for the dns-cloudflare and dns-digitalocean plugins) plugins are
now included in the Certbot ebuild, selectable by USE flags.
Therefore, these separate plugins are no longer necessary. Only the
certbot-dns-cloudflare and certbot-dns-digitalocean will be kept as
they are not included in the official Certbot ebuild.